he cambiado el constructor de Balsa para que admita double en tiempo

+++ src/Balsa.java
@@ -1,9 +1,9 @@
 public class Balsa {
 
     private int capacidad;
-    private int tiempo;
+    private double tiempo;
 
-    public Balsa(int capacidad, int tiempo) {
+    public Balsa(int capacidad, double tiempo) {
         this.capacidad = capacidad;
         this.tiempo = tiempo;
     }
